==> notebooks/0.0-introduction.ipynb <==
<div style="text-align: center; background-color: #0A6EBD; font-family: 'Trebuchet MS', Arial, sans-serif; color: white; padding: 20px; font-size: 40px; font-weight: bold; border-radius: 0 0 0 0; box-shadow: 0px 6px 8px rgba(0, 0, 0, 0.2);">
  Final Project - Introduction to Data Science
    <div style="text-align: center; background-color: #0A6EBD; font-family: 'Trebuchet MS', Arial, sans-serif; color: white; padding: 20px; font-size: 40px; font-weight: bold; border-radius: 0 0 0 0; box-shadow: 0px 6px 8px rgba(0, 0, 0, 0.2);">
    @ FIT-HCMUS, VNU-HCM 📌
</div>

<div style="text-align: center; background-color: #5A96E3; font-family: 'Trebuchet MS', Arial, sans-serif; color: white; padding: 20px; font-size: 40px; font-weight: bold; border-radius: 0 0 0 0; box-shadow: 0px 6px 8px rgba(0, 0, 0, 0.2);">
  Stage 00 - A brief introduction to course assignment 2📌
</div>

## Workflow stages

Generic workflow stages, however there are use cases with exceptions.
1. Overview: 
    - Project title.
    - What are the goals of the project (question or problem definition)?
    - What is(are) dataset(s) ?
    - Describe your team members.
2. Data overview
    - What is the source of data? Public? Crawled?
    - What is the size of data?
    - Brief introduction of data? What is the mean of each data properties?
3. Data preparation and cleaning
    - Acquire training and testing data.
    - Wrangle, prepare, cleanse the data.
    - Missing value treatment.
    - Outlier detection and treatment.
    - Feature engineering.
4. Exploratory data analysis: Analyze, identify patterns, and explore the data.
    - Univariate analysis.
    - Bivariate analysis.
    - Multivariate analysis.
5. Data visualization
    - Graphs, plots.
    - Describe insights derivered from the visuals (combine with outside information from dataset that you found).
6. Hypothesis testing and insights
    - Statistical tests used
    - Findings and insights
7. Final report and presentation
    - Key findings.
    - What are limitations of the analysis?
    - Propose recommendations for the next steps.
8. References and data sources
    - Structure dataset(s) and export links.
    - Listing research papers or articles in reference section.
9. Project structurization
    - Notebooks source codes
    - Python source codes
    - Model binary files
    - Data files
    - Presentation or report files.

## Workflow goals

The data science solutions workflow solves for <font color=lightgreen>seven major goals (7C)</font>.

<font color=lightgreen>**Classifying**</font>: Classifying involves classifying or categorizing data into different classes or categories. It's often used in classification problems where you aim to assign a label or category to data points. We may also want to understand the implications or correlation of different classes with our solution goal.

<font color=lightgreen>**Correlating**</font>: Considering the available features within the training dataset, one wants to identify and analyse the relationships between them, and then determine which features within the dataset significantly contribute to our solution goal. Generally, correlating refers to the process of identifying and analysing relationships or correlations between different variables in the data. This step helps understand how variables are related to each other and is often part of exploratory data analysis (EDA). Correlating certain features may help in creating, completing, or correcting features.

<font color=lightgreen>**Converting**</font>: In the data modelling stage, one should prepare the data to satisfy the requirements of learning algorithms. Converting data may involve changing the format or data type of variables to make them compatible with the analysis or modelling techniques. For example, converting text data into numerical features using techniques like one-hot encoding.

<font color=lightgreen>**Completing**</font>: Due to many reasons, such as weak data collection techniques or low-quality data resources, datasets often have the imcompleteness characteristic. Completing the data involves addressing missing values in the dataset. Missing data can impact the quality of analysis and modelling, so it's important to handle missing values through imputation or other techniques.

<font color=lightgreen>**Correcting**</font>: We may also search for errors or wrong values within features in the given training dataset and either confirm these values or reject the samples that include the errors. The process of finding and repairing faults or inconsistencies in a dataset is known as data rectification. Looking for outliers in our samples or attributes is one way. Another alternative is to eliminate a characteristic if it does not contribute to the analysis or has the potential to significantly affect the results.

<font color=lightgreen>**Creating**</font>: Feature engineering or simply creating is a critical phase in the data science pipeline for creating new features or variables. It entails creating new variables or modifying existing ones in order to improve the performance of machine learning models.

<font color=lightgreen>**Charting**</font>: Charting and data visualisation are critical components of the data science workflow. It entails producing charts, graphs, and other visual representations of data in order to better analyse patterns, trends, and insights. Data visualisation may help with communication and narrative.

## Assignment goals

This assignment is an extended version of Assignment 1. The main purpose here is to make an exploratory analysis of the demographic and socio-economic data provided by the World Bank. Thus, this assignment will lead a brief introduction to the data science process for students who have enrolled in the Introduction to Data Science course 2023 at FIT-HCM, VNU-HCM. After success in crawling data using the Web API (http://api.worldbank.org/v2/country/all/indicator/SP.POP.TOTL) in Assignment 1, Once again, you will have to crawl data by using this API, but on another scale.

1. You will learn how to setup a data science project. How do you structure it as clearly as possible?
2. You will learn basic techniques in data pre-processing, basic exploratory data analysis (EDA) from making questions and finding the answer from insights developed from statistical methods, and simple visualisation methods.
3. You will learn to set up a report clearly.

## Set up your coding environment with Anaconda

First, you should [download and install Anaconda](https://www.anaconda.com/download) for your operating systems. Then, starting install your development environment using ```requirements.txt```. 

**Note**: Suppose that you want to use Python 3.9, I built this requirement packages list base on Python 3.9. So if you want to install these packages in other Python version, please check all packages again via command ```conda search -c conda-forge -f  <package name>``` to verify the version of each package and change it in ```requirements.txt``` file. 

```bash
# create conda virtual environment
conda create --name min_ds-env python=3.9 -y

# activate created conda virtual environment
conda activate min_ds-env

# install dependencies
pip install -r requirements.txt
```

## How to complete this assignment?

You will do your assignment directly in this notebook. First, fill in your name and ID at the beginning of the file. In the notebook, fill in places that say:
```python
#TODO
```

During your work, you can print out the result, create more cells to test, or create more functions to handle things. Please note that <font color=red>you are not allowed to delete or modify my code cells</font> (except in the case that mentioned above). Let remove `raise NotImplementedError("not implement")` when running code.

Always press `Ctrl + S` in order to save your work.

**Notes:** 

    *  Copy means zero
    *  You have to submit your work on time. No exception
    *  Any questions about your grade are based on the assignment submitted on Moodle
    *  Wrong submission takes you -2 points

## How to submit this assignment?

When grading your assignment, I will choose `Kernel` - `Restart Kernel & Run All Cells` in order to restart the kernel and run all cells in your notebook. Therefore, you should do that before submitting to ensure that the outputs are all as expected.

After that, you make a submited direction as follow:

- Folder `StudentCode` (e.g. If your student code is 1234567, then your folder is `1234567`)

Finally, you compress your folder (`StudentCode`) and submit on Moodle. **The extension of the file is nothing else but `.zip`.**

<font color=red>Please strictly follow the submission rules.</font>

## Laboratory structure explanation

This lab project structure is built under the [Cookiecutter Data Science](https://github.com/drivendata/cookiecutter-data-science) - A logical, reasonably standardized, but flexible project structure for doing and sharing data science work. 

```
├── LICENSE-CC-BY-NC-ND-4.0.md
├── min_ds-env.yml
├── Makefile                 <- Makefile with commands like `make data` or `make train`
├── README.md                <- The top-level README for developers using this project.
├── data
│   ├── external             <- Data from third party sources.
│   ├── interim              <- Intermediate data that has been transformed.
│   ├── processed            <- The final, canonical data sets for modeling.
│   └── raw                  <- The original, immutable data dump.
│
│
├── models                   <- Trained and serialized models, model predictions, or model summaries
│
├── notebooks                <- Jupyter notebooks. Naming convention is a number (for ordering),
│                               the creator's initials, and a short `-` delimited description, e.g.
│                               `1.0-initial-data-exploration`.
│
├── references                <- Data dictionaries, manuals, and all other explanatory materials.
│
├── reports                   <- Generated analysis as HTML, PDF, LaTeX, etc. or your written by template such as https://github.com/Wandmalfarbe/pandoc-latex-template
│   └── figures               <- Generated graphics and figures to be used in reporting
│
├── requirements.txt          <- The requirements file for reproducing the analysis environment
│
├── src                       <- Source code for use in this project.
│   ├── __init__.py           <- Makes src a Python module
│   │
│   ├── data_module           <- Scripts to download or generate data
│   │   └── make_dataset.py
│   │
│   ├── features_module       <- Scripts to turn raw data into features for modeling
│   │   └── build_features.py
│   │
│   ├── models_module         <- Scripts to train models and then use trained models to make
│   │   │                        predictions
│   │   ├── predict_model.py
│   │   └── train_model.py
│   │
│   └── visualization_module  <- Scripts to create exploratory and results oriented visualizations
│       └── visualize.py
```

## Contact information

If you have any concerns with this assignment, please write an email and send it to class lecturer or laboratory instructor in order to get useful advices
- Class lecturer: **Le Ngoc Thanh** (lnthanh@fit.hcmus.edu.vn)
- Course laboratory instructor: **Le Nhut Nam** (lenam.fithcmus@gmail.com)

## LICENSE

Lab 02 - Introduction To Data Science/ CSC14119 © 2023 by Department of Computer Science, Faculty of Information Technology, University of Science, Vietnam National University, Ho Chi Minh City is licensed under Attribution-NonCommercial-NoDerivatives 4.0 International

Please read the file ```LICENSE-CC-BY-NC-ND-4.0.md``` for more understanding.
