==> notebooks/2.0-data-preprocessing.ipynb <==
<div style="text-align: center; background-color: skyblue; font-family: 'Times New Roman', serif; color: white; padding: 20px; font-size: 36px; font-weight: bold; border-radius: 15px; box-shadow: 0px 6px 15px rgba(0, 0, 0, 0.2); transition: all 0.3s ease;">
    INTRODUCTION TO DATA SCIENCE<br>
    @ FIT-HCMUS, VNU-HCM<br>
    FINAL PROJECT 🏫
</div>
<style>
    div:hover {
        transform: scale(1.05);
        box-shadow: 0px 10px 20px rgba(0, 0, 0, 0.3);
    }
</style>

<div style="text-align: center; background-color: #ADD8E6; font-family: 'Times New Roman', serif; color: white; padding: 20px; font-size: 30px; font-weight: bold; border-radius: 10px; box-shadow: 0px 6px 15px rgba(0, 0, 0, 0.2); transition: all 0.3s ease;">
    THE ANALYSIS OF THE MOBILE PHONE MARKET AT MOBILE CITY STORE
</div>
<style>
    div:hover {
        transform: scale(1.05);
        box-shadow: 0px 10px 20px rgba(0, 0, 0, 0.3);
    }
</style>


---

<div style="text-align: center; background-color: #ADD8E6; font-family: 'Times New Roman', serif; color: white; padding: 25px; font-size: 20px; font-weight: bold; border-radius: 10px; box-shadow: 0px 6px 15px rgba(0, 0, 0, 0.2); transition: all 0.3s ease;">
    DATA PREPROCESSING
</div>
<style>
    div:hover {
        opacity: 0.8;
    }
</style>

**Người thực hiện**: Bùi Đình Bảo

**MSSV**: 21120201


---

# Thư viện

import numpy as np
import pandas as pd

import ast
import re

pd.set_option('display.max_columns', None)
pd.set_option('display.max_rows', None)

data = pd.read_csv("../data/raw_data.csv")
data.drop_duplicates(inplace=True)
data.head()

# Đặt lại tên thuộc tính

data.columns = ['ten', 'loai_dien_thoai', 'mau_sac_phien_ban_bo_nho_gia_moi_gia_cu', 'thoi_gian_bao_hanh', 'thong_so_ky_thuat', 'danh_gia', 'so_luong_binh_luan', 'duong_dan']
data.head()

> Xóa các dòng có chứa xe ô tô điện?

data = data[data["ten"]!="Điện thoại Xiaomi SU7 (Ô tô điện chạy 800km/1 lần sạc)"].reset_index(drop=True)

> Xóa các dòng có chứa Apple Watch

data = data[data["loai_dien_thoai"]!="Apple Watch"]

> Xóa các dòng có chứa máy chơi game PC

data = data[data["loai_dien_thoai"]!="Máy Chơi Game PC"]

# Trích xuất thuộc tính cơ bản

## Thuộc tính: Màu sắc, Phiên bản bộ nhớ, Giá mới, Giá cũ

data["mau_sac_phien_ban_bo_nho_gia_moi_gia_cu"] = data["mau_sac_phien_ban_bo_nho_gia_moi_gia_cu"].apply(ast.literal_eval)
data = data.explode("mau_sac_phien_ban_bo_nho_gia_moi_gia_cu").reset_index(drop=True)
df_expanded = pd.DataFrame(data['mau_sac_phien_ban_bo_nho_gia_moi_gia_cu'].tolist(), columns=['mau_sac', 'phien_ban_bo_nho', 'gia_moi', 'gia_cu'])
data = pd.concat([data, df_expanded], axis=1).drop("mau_sac_phien_ban_bo_nho_gia_moi_gia_cu", axis=1)

## Thuộc tính: Màn hình, Hệ điều hành, Camera sau, Camera trước, CPU, RAM, Bộ nhớ trong, Thẻ SIM, Dung lượng pin, Thiết kế

explode_specifications = data.thong_so_ky_thuat.apply(ast.literal_eval).apply(pd.Series)
explode_specifications.columns = ['man_hinh', 'he_dieu_hanh', 'camera_sau', 'camera_truoc', 'cpu', 'ram', 'bo_nho_trong', 'the_sim', 'dung_luong_pin', 'thiet_ke']
data = pd.concat([data, explode_specifications], axis=1)
data.drop("thong_so_ky_thuat", axis=1, inplace=True)

## Thuộc tính: Hãng điện thoại

def get_brand(s):
    if type(s)!=str:
        return None
    s = s.replace("Điện thoại", "")
    s = s.strip()
    res = s.split(" ")[0].strip()
    if res == "OPPO":
        return "Oppo"
    if "Bkav" in res:
        return "Bphone"
    if res == "Thành":
        return "Thành Hưng"
    return res

data["hang_dien_thoai"] = data["ten"].apply(get_brand)

## Thuộc tính: Có phải là điện thoại cũ hay không?

def isOld(s):
    if type(s)!=str:
        return None
    s = s.lower()
    if "cũ" in s:
        return True
    return False

data["old1"] = data["ten"].apply(isOld)
data["old2"] = data["loai_dien_thoai"].apply(isOld)

data["la_dien_thoai_cu"] = data["old1"] | data["old2"]
data.drop(columns=["old1", "old2"], inplace=True)

# Chuyển đổi định dạng dữ liệu và trích xuất thêm thuộc tính

## Thuộc tính: Giá mới, Giá cũ

def cast_price(s):
    if type(s)!=str:
        return np.nan
    s = s.replace(".", "")
    numbers = re.findall(r'\d+', s)
    if len(numbers) == 0:
        return np.nan
    return int(numbers[0])
data["gia_moi"] = data["gia_moi"].apply(cast_price)
data["gia_cu"] = data["gia_cu"].apply(cast_price)

## Thuộc tính: Thời gian bảo hành

def cast_warranty(s):
    if type(s)!=str:
        return np.nan
    numbers = re.findall(r'\d+', s)
    if len(numbers)==0:
        return np.nan
    if len(numbers)==1:
        if numbers[0]==30:
            return 1
        return int(numbers[0])
    if numbers[0]==6 and numbers[1]==12:
        return 12
    if numbers[0]==12 and numbers[1]==12:
        return 24
    return int(numbers[0])
data["thoi_gian_bao_hanh"] = data["thoi_gian_bao_hanh"].apply(cast_warranty)

## Thuộc tính: Đánh giá

def cast_rating(s):
    if type(s)!=str:
        return np.nan
    numbers = re.findall(r"-?\d+(?:\.\d+)?", s)
    if len(numbers)==0:
        return np.nan
    return float(numbers[0])
data["danh_gia"] = data["danh_gia"].apply(cast_rating)

## Thuộc tính: Phiên bản bộ nhớ

def cast_memory_version(s):
    if type(s)!=str:
        return [np.nan, np.nan]
    numbers = re.findall(r"-?\d+(?:\.\d+)?", s)
    if len(numbers)==0:
        return [np.nan, np.nan]
    if len(numbers)==1:
        if abs(float(numbers[0]))<4:
            return [np.nan, abs(float(numbers[0])*1024)]
        return [np.nan, abs(float(numbers[0]))]
    if abs(float(numbers[1]))<4:
        return [abs(float(numbers[0])), abs(float(numbers[1])*1024)]
    return [abs(float(numbers[0])), abs(float(numbers[1]))]
data["phien_ban_bo_nho"] = data["phien_ban_bo_nho"].apply(cast_memory_version)
df_expanded = pd.DataFrame(data['phien_ban_bo_nho'].tolist(), columns=['_ram', '_bo_nho_trong'])
data = pd.concat([data, df_expanded], axis=1).drop("phien_ban_bo_nho", axis=1)

## Thuộc tính: RAM, Bộ nhớ trong

def firstNumberRAM(s):
    if type(s)!=str:
        return np.nan
    numbers = re.findall(r"-?\d+(?:\.\d+)?", s)
    if len(numbers)==0:
        return np.nan
    res = float(numbers[0])
    return res if res < 100 else res*1.0/1024

def firstNumberROM(s):
    if type(s)!=str:
        return np.nan
    numbers = re.findall(r"-?\d+(?:\.\d+)?", s)
    if len(numbers)==0:
        return np.nan
    res = float(numbers[0])
    if res < 2:
        res = res*1024

data["ram"] = data["ram"].apply(firstNumberRAM)
data["bo_nho_trong"] = data["bo_nho_trong"].apply(firstNumberROM)
data["ram"] = data["_ram"].fillna(data["ram"])
data["bo_nho_trong"] = data["_bo_nho_trong"].fillna(data["bo_nho_trong"])
data = data.drop(columns=["_ram", "_bo_nho_trong"])

data[data["bo_nho_trong"]<=8]

## Thuộc tính: Kích thước màn hình, Tần số quét, Độ sáng màn hình, Loại màn hình

> Bao gồm trích xuất thuộc tính

def extractInches(s):
    if type(s)!=str:
        return np.nan
    s = s.lower()
    # match = re.search(r'(\d+(\.\d+)?)\s*inches', s)
    match = re.search(r'(\d+[.,]?\d*)\s*inches', s)
    result = float(match.group(1).replace(",", ".")) if match else np.nan
    return result
def extractHz(s):
    if type(s)!=str:
        return np.nan
    s = s.lower()
    match = re.search(r'(\d+(\.\d+)?)\s*hz', s)
    result = float(match.group(1)) if match else np.nan
    return result
def extractNits(s):
    if type(s)!=str:
        return np.nan
    s = s.lower()
    match = re.search(r'(\d+(\.\d+)?)\s*nits', s)
    result = float(match.group(1)) if match else np.nan
    if result < 50:
        result = result*1000
    return result
def extractType(s):
    if type(s)!=str:
        return np.nan
    s = s.upper()
    if "AMOLED" in s:
        return "AMOLED"
    if "OLED" in s:
        return "OLED"
    if "LCD" in s:
        return "LCD"
    
data["kich_thuoc_man_hinh"] = data["man_hinh"].apply(extractInches)
print(data.kich_thuoc_man_hinh.isna().sum())
data["tan_so_quet"] = data["man_hinh"].apply(extractHz)
print(data.tan_so_quet.isna().sum())
data["do_sang_man_hinh"] = data["man_hinh"].apply(extractNits)
print(data.do_sang_man_hinh.isna().sum())
data["loai_man_hinh"] = data["man_hinh"].apply(extractType)
print(data.loai_man_hinh.isna().sum())
### Còn độ phân giải ...

data = data.drop(columns=["man_hinh"])

## Thuộc tính: Số thẻ SIM

def extractSim(s):
    if type(s)!=str:
        return np.nan
    s = s.lower()
    if "2" in s or "dual" in s or "kép" in s:
        return 2
    if "không" in s:
        return 0
    return 1
data["so_the_sim"] = data["the_sim"].apply(extractSim)
data = data.drop(columns=["the_sim"])

## Thuộc tính: Loại pin, Dung lượng pin, Công suất sạc

> Bao gồm trích xuất thuộc tính

def batteryType(s):
    if type(s)!=str:
        return ""
    s = s.lower()
    if "li-po" in s:
        return "Li-Po"
    if "li-ion" in s:
        return "Li-Ion"
    if "si/c" in s:
        return "Si/C"
    return ""
def batteryCapacity(s):
    if type(s)!=str:
        return np.nan
    s = s.lower()
    match = re.search(r'(\d+(\.\d+)?)\s*mah', s)
    result = float(match.group(1)) if match else np.nan
    return result
def chargingCapacity(s):
    if type(s)!=str:
        return np.nan
    s = s.lower()
    match = re.search(r'(\d+(\.\d+)?)\s*w', s)
    result = float(match.group(1)) if match else np.nan
    return result
data["pin"] = data["dung_luong_pin"]
data["loai_pin"] = data["pin"].apply(batteryType)
data["dung_luong_pin"] = data["pin"].apply(batteryCapacity)
data["cong_suat_sac"] = data["pin"].apply(chargingCapacity)
data = data.drop(columns=["pin"])

## Thuộc tính: Độ phân giải camera trước, Độ phân giải camera sau

def extractMP(s):
    if type(s)!=str:
        return []
    s = s.upper()
    matches = re.findall(r'(\d+(\.\d+)?)\s*MP', s)
    result = [float(match[0]) for match in matches]
    return result
data["do_phan_giai_cam_sau"] = data["camera_sau"].apply(extractMP)
data["do_phan_giai_cam_truoc"] = data["camera_truoc"].apply(extractMP)
data = data.drop(columns=["camera_sau", "camera_truoc"])

## Thuộc tính: Hệ điều hành

def cast_OS(s):
    if type(s)!=str:
        return "Khác"
    s = s.lower()
    numbers = re.findall(r'\d+', s)
    if len(numbers)==0:
        return "Khác"
    if "android" in s:
        return "Android " + str(int(numbers[0]))
    if "ios" in s:
        return "iOS " + str(int(numbers[0]))
    return "Khác"
data["he_dieu_hanh"] = data["he_dieu_hanh"].apply(cast_OS)

data.info()
data.head()

# Xử lý dữ liệu thiếu

> Không xử lý ở đây mà sẽ xử lý trước khi trực quan hóa hoặc xây dựng mô hình ở những notebook tiếp theo.

## Thời gian bảo hành

data["thoi_gian_bao_hanh"] = data.thoi_gian_bao_hanh.fillna(0)

## Màu sắc

data["mau_sac"] = data["mau_sac"].replace("", None)

---

data.info()
data.head()

data.to_csv("../data/processed_data.csv", index=True)
